requirements.txt
docker>=3.0.1
PyYAML>=3.12